requirements/requirements-enterprise.txt
# Enterprise Business Intelligence Platform Dependencies
# Extends base requirements.txt with enterprise features

# Base requirements (install requirements.txt first)
-r requirements.txt

# Enterprise API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Async HTTP Client
httpx>=0.25.0
aiohttp>=3.9.0

# Security & Encryption
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Database & Caching
redis>=5.0.0
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.0
alembic>=1.13.0

# Monitoring & Observability
psutil>=5.9.0
prometheus-client>=0.19.0
structlog>=23.2.0

# Task Queue & Background Jobs
celery>=5.3.0
kombu>=5.3.0

# Data Validation & Serialization
marshmallow>=3.20.0
cerberus>=1.3.4

# Enterprise Integrations
requests-oauthlib>=1.3.1
msal>=1.25.0
simple-salesforce>=1.12.0

# Testing (Development)
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
httpx>=0.25.0
factory-boy>=3.3.0

# Performance & Profiling
memory-profiler>=0.61.0
py-spy>=0.3.14

# Configuration Management
python-dotenv>=1.0.0
pydantic-settings>=2.1.0

# Logging & Monitoring
loguru>=0.7.0
sentry-sdk[fastapi]>=1.38.0

# API Documentation
fastapi[all]>=0.104.0

# Enterprise Security
bandit>=1.7.5
safety>=2.3.0

# Load Testing
locust>=2.17.0
